requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
click==8.3.1
fastapi==0.128.0
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0

typer~=0.9.0
rich~=14.2.0
pip~=25.3
attrs~=25.4.0
distro~=1.9.0
wheel~=0.45.1
protobuf~=6.33.1
cryptography~=41.0.4
pillow~=12.0.0
filelock~=3.19.1
sniffio~=1.3.1
email-validator~=2.3.0
requests~=2.31.0
itsdangerous~=2.2.0
Jinja2~=3.1.6
httpx~=0.28.1
SQLAlchemy~=2.0.45
passlib~=1.7.4
python-jose~=3.5.0
tensorflow~=2.20.0
python-dotenv==1.0.1
pytest
pytest-cov